Create price_calculator
There is a total of only 4 products you sell. Each one has a different price but they are all taxed equally. The goal is to give back the total price including the tax. In this town, the tax is 6% on top of the total. There will be a main function, another function to add things up, and a last one to add the taxes. Main will return the total.